backend/pyproject.toml
[project]
name = "research-project"
version = "0.1.0"
description = ""
authors = [
    {name = "OkunrinmetaWebDevelopment",email = "c.okunrinmetrix@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.15"
dependencies = [
    "pydantic (>=2.10.6,<3.0.0)",
    "sqlalchemy (>=2.0.39,<3.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "python-settings (>=0.2.3,<0.3.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "python-jose (>=3.4.0,<4.0.0)",
    "cryptography (>=44.0.2,<45.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "pyisemail (>=2.0.1,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "cryptocode (>=0.1,<0.2)",
    "boto3 (==1.35.24)",
    "pillow (==10.4.0)",
    "clicksend-client (>=5.0.78,<6.0.0)",
    "bcrypt (==3.2.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "google-auth (>=2.40.2,<3.0.0)",
    "google-api-python-client (>=2.170.0,<3.0.0)",
    "google-auth-oauthlib (>=1.2.2,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "trafilatura (>=1.12.2,<2.0.0)",
    "supabase (>=2.10.0,<3.0.0)",
    "fastembed>=0.7.3",
    "langchain>=0.3.27",
    "langchain-anthropic>=0.3.21",
    "langchain-huggingface>=0.3.1",
    "langchain-ollama>=0.3.9",
    "langchain-openai>=0.3.34",
    "langchain-together>=0.3.1",
    "faiss-cpu>=1.12.0",
    "numpy>=2.3.3",
    "redis",
    "langchain-community (>=0.3.30,<0.4.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"